Ajout de tous les endpoints de rentals + optimize import on the fly

+++ ChaTopBackend/src/main/java/com/openclassrooms/repository/DBUserRepository.java
@@ -3,6 +3,6 @@ package com.openclassrooms.repository;
 import com.openclassrooms.model.DBUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DBUserRepository extends CrudRepository<DBUser, Integer> {
+public interface DBUserRepository extends CrudRepository<DBUser, Long> {
   DBUser findByEmail(String email);
 }
